requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==2.2.8
django-cors-headers==2.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
djongo==1.3.6
PyJWT==2.4.0
pymongo==3.12.1
pytz==2022.1
sqlparse==0.2.4
